backend/services/synthetic-tests/requirements.txt
# Synthetic Test Framework Requirements

# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.13.0

# HTTP client
aiohttp==3.9.1

# Google Cloud Pub/Sub
google-cloud-pubsub==2.18.4

# Redis for caching
redis==5.0.1
aioredis==2.0.1

# Background tasks
celery==5.3.4
redis==5.0.1

# JSON schema validation
jsonschema==4.20.0

# Expression evaluation
simpleeval==0.9.13

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Logging and monitoring
structlog==23.2.0
